# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for personal website with CV page and AI Blog",
        "title": "Personal Web Platform API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/admin/profile": {
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Update CV/profile information (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Update profile information",
                "parameters": [
                    {
                        "description": "Profile update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_transport_http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_transport_http.ProfileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/profile": {
            "get": {
                "description": "Get public CV/profile information (accessible to everyone)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Get profile information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_transport_http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_transport_http.ProfileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the service is alive (liveness probe)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http.HealthResponse"
                        }
                    }
                }
            }
        },
        "/ready": {
            "get": {
                "description": "Check if the service is ready to accept traffic (readiness probe)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Readiness check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http.ReadinessResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http.ReadinessResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_transport_http.ContactsResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "github": {
                    "type": "string",
                    "example": "https://github.com/johndoe"
                },
                "linkedin": {
                    "type": "string",
                    "example": "https://linkedin.com/in/johndoe"
                }
            }
        },
        "internal_transport_http.ErrorData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_transport_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal_transport_http.ErrorData"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "internal_transport_http.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_transport_http.MetaData": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "internal_transport_http.ProfileResponse": {
            "type": "object",
            "properties": {
                "activity": {
                    "type": "string",
                    "example": "Building awesome products"
                },
                "contacts": {
                    "$ref": "#/definitions/internal_transport_http.ContactsResponse"
                },
                "description": {
                    "type": "string",
                    "example": "Senior Software Engineer"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "photo_url": {
                    "type": "string",
                    "example": "https://example.com/photo.jpg"
                }
            }
        },
        "internal_transport_http.ReadinessResponse": {
            "type": "object",
            "properties": {
                "checks": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "ready": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "internal_transport_http.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/internal_transport_http.ErrorData"
                },
                "meta": {
                    "$ref": "#/definitions/internal_transport_http.MetaData"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "internal_transport_http.UpdateProfileRequest": {
            "type": "object",
            "properties": {
                "activity": {
                    "type": "string",
                    "example": "Building products"
                },
                "description": {
                    "type": "string",
                    "example": "Senior Engineer"
                },
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "github": {
                    "type": "string",
                    "example": "https://github.com/johndoe"
                },
                "linkedin": {
                    "type": "string",
                    "example": "https://linkedin.com/in/johndoe"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "photo_url": {
                    "type": "string",
                    "example": "https://example.com/photo.jpg"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "Session cookie authentication",
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}